Doc finale Aurélien
Derniers changements du code :

~ fenetre de résultat non modal -> ca bloquait le serveur !!! du coup toutes les créatures mourraient chez les clients

- les cheats

~ amélioration du panel de lancement de créatures (meilleure màj)

~ les pieces d'or sont maintenant en double -> on est obligé car les revenus sont en double... c'est nettement plus précis!

+++ src/models/joueurs/Joueur.java
@@ -39,8 +39,10 @@ public class Joueur
      * <br>
      * Cette variable fluctue en fonction des creatures tuees et de 
      * l'achat et vente des tours.
+     * 
+     * Elle est en double pour gérer les revenus flottant (voir gestionnaireDeRevenus)
      */
-    private int nbPiecesDOr = 0;
+    private double nbPiecesDOr = 0;
     
     /**
      * Equipe du joueur
@@ -132,7 +134,7 @@ public class Joueur
      * 
      * @param nbPiecesDOr le nouveau nombre de pieces d'or
      */
-    public void setNbPiecesDOr(int nbPiecesDOr)
+    public void setNbPiecesDOr(double nbPiecesDOr)
     {
         this.nbPiecesDOr = nbPiecesDOr;
         
@@ -145,7 +147,7 @@ public class Joueur
      * 
      * @return le nombre de pieces d'or du joueur
      */
-    public int getNbPiecesDOr()
+    public double getNbPiecesDOr()
     {
         return nbPiecesDOr;
     }
@@ -308,7 +310,7 @@ public class Joueur
      */
     public synchronized void donnerRevenu(long temps)
     {
-        setNbPiecesDOr((int) (getNbPiecesDOr() + revenu * (temps / 1000.0)));
+        setNbPiecesDOr(getNbPiecesDOr() + revenu * (temps / 1000.0));
     }
